Move `LineSentence` and `Text8Corpus` to `textcorpus` module. Rename `textcorpus.walk` to `walk_with_depth` and move to `utils` module. Update tests and other referencing modules to adjust to the moves, resolving some circular references that arose in the process.
