+++ gensim/models/phrases.py
@@ -461,7 +461,8 @@ if __name__ == '__main__':
     infile = sys.argv[1]
 
     from gensim.models import Phrases  # for pickle
-    from gensim.models.word2vec import Text8Corpus
+    from gensim.corpora.textcorpus import Text8Corpus
+
     sentences = Text8Corpus(infile)
 
     # test_doc = LineSentence('test/test_data/testcorpus.txt')
