+++ gensim/test/test_textcorpus.py
@@ -10,14 +10,17 @@ Automated tests for checking textcorpus I/O formats.
 
 from __future__ import unicode_literals
 
+import bz2
 import codecs
 import logging
 import os
 import tempfile
 import unittest
 
+from gensim import utils
 from gensim.corpora import textcorpus
 from gensim.test.test_corpora import CorpusTestCase, datapath
+from gensim.test.test_word2vec import datapath
 
 
 def split_tokenize(text):
@@ -270,6 +273,49 @@ class TestTextDirectoryCorpus(unittest.TestCase):
 
         corpus.pattern = "4.*"
         self.assertEqual(expected[-1], base_names[-1])
+
+
+class TestWord2VecSentenceIterators(unittest.TestCase):
+    def testLineSentenceWorksWithFilename(self):
+        """Does LineSentence work with a filename argument?"""
+        with utils.smart_open(datapath('lee_background.cor')) as orig:
+            sentences = textcorpus.LineSentence(datapath('lee_background.cor'))
+            for words in sentences:
+                self.assertEqual(words, utils.to_unicode(orig.readline()).split())
+
+    def testLineSentenceWorksWithCompressedFile(self):
+        """Does LineSentence work with a compressed file object argument?"""
+        with utils.smart_open(datapath('head500.noblanks.cor')) as orig:
+            sentences = textcorpus.LineSentence(bz2.BZ2File(datapath('head500.noblanks.cor.bz2')))
+            for words in sentences:
+                self.assertEqual(words, utils.to_unicode(orig.readline()).split())
+
+    def testLineSentenceWorksWithNormalFile(self):
+        """Does LineSentence work with a file object argument, rather than filename?"""
+        with utils.smart_open(datapath('head500.noblanks.cor')) as orig:
+            with utils.smart_open(datapath('head500.noblanks.cor')) as fin:
+                sentences = textcorpus.LineSentence(fin)
+                for words in sentences:
+                    self.assertEqual(words, utils.to_unicode(orig.readline()).split())
+
+    def testPathLineSentences(self):
+        """Does PathLineSentences work with a path argument?"""
+        with utils.smart_open(os.path.join(datapath('PathLineSentences'), '1.txt')) as orig1, \
+                utils.smart_open(os.path.join(datapath('PathLineSentences'), '2.txt.bz2')) as orig2:
+            sentences = word2vec.PathLineSentences(datapath('PathLineSentences'))
+            orig = orig1.readlines() + orig2.readlines()
+            orig_counter = 0  # to go through orig while matching PathLineSentences
+            for words in sentences:
+                self.assertEqual(words, utils.to_unicode(orig[orig_counter]).split())
+                orig_counter += 1
+
+    def testPathLineSentencesOneFile(self):
+        """Does PathLineSentences work with a single file argument?"""
+        test_file = os.path.join(datapath('PathLineSentences'), '1.txt')
+        with utils.smart_open(test_file) as orig:
+            sentences = word2vec.PathLineSentences(test_file)
+            for words in sentences:
+                self.assertEqual(words, utils.to_unicode(orig.readline()).split())
 
 
 if __name__ == '__main__':
